% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoplanets.R
\name{exo_column_names}
\alias{exo_column_names}
\title{Access column names for each table}
\usage{
exo_column_names(table = "exoplanets", cols = "default",
  progress = TRUE, col_spec = FALSE)
}
\arguments{
\item{table}{The name of the table, see \code{exo_tables} for all available
table names.}

\item{cols}{Either "default" for default columns, "all" for all columns,
defaults to "default".}

\item{progress}{If FALSE, suppresses progress of request.}

\item{col_spec}{If FALSE, suppresses column specification message from
\code{readr}.}
}
\value{
A character vector containing column names for the respective table
}
\description{
Simply pull column names for a specified table. You can either
pull the default columns assigned to a table or all columns.
}
\examples{
x <- exo_column_names(
  table = "cumulative",
  cols = "default",
  progress = FALSE
  )
str(x)
}
