% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exo_columns.R
\name{exo_columns}
\alias{exo_columns}
\title{Access column names for each table}
\usage{
exo_columns(table = "exoplanets", cols = "default", quiet = FALSE)
}
\arguments{
\item{table}{The name of the table, see \code{exo_tables} for all available
table names.}

\item{cols}{Either "default" for default columns, "all" for all columns,
defaults to "default".}

\item{quiet}{If FALSE, suppress query message.}
}
\value{
A character vector containing column names for the respective table
}
\description{
Simply pull column names for a specified table. You can either
pull the default columns assigned to a table or all columns.
}
