% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoplanets.R
\name{exo_kelt}
\alias{exo_kelt}
\title{KELT light curves (not including KELT Praesepe data)}
\usage{
exo_kelt(kelt_field = NULL, progress = TRUE)
}
\arguments{
\item{kelt_field}{KELT field of observation. Fields available are N02, N04,
N06, N08, N10 and N12.}

\item{progress}{If FALSE, suppresses progress of request (unix OS only).}
}
\description{
The NASA Exoplanet Archive provides access to KELT time series through a
search interface and the application programming interface (API). The
following table lists all the data columns in the KELT Time Series table
(kelttimeseries) that can be returned through the API. See the KELT page
for more information on the KELT data products available at the NASA
Exoplanet Archive.
}
