% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exo_summary.R
\name{exo_summary}
\alias{exo_summary}
\title{Exoplanet Summary}
\usage{
exo_summary(quiet = FALSE)
}
\arguments{
\item{quiet}{If FALSE, suppress query message.}
}
\value{
An object of class \code{data.frame}.
}
\description{
A list of data summarising the exoplanet database. The goal of
this function is to provide the information found here
<\url{https://exoplanetarchive.ipac.caltech.edu/docs/counts_detail.html}>.
}
