% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoplanets.R
\docType{data}
\name{exo_tables}
\alias{exo_tables}
\title{All Available Tables}
\format{An object of class \code{list} of length 32.}
\usage{
exo_tables
}
\description{
This dataset provides a list of all available tables and
their names, convenient for calling on tables when using \code{exo()}.
Please note that a few tables require additional parameters. For example,
the \code{keplertimeseries} table requires either the quarter or a Kepler
ID must be specified, e.g. \code{exo("keplertimeseries&quarter=14")}.
This requirements only occurs for a small number of tables, mostly time
series.
}
\keyword{datasets}
