% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exo_wasp.R
\name{exo_wasp}
\alias{exo_wasp}
\title{SuperWASP Time Series Table}
\usage{
exo_wasp(tile = NULL, sourceid = NULL, quiet = FALSE)
}
\arguments{
\item{tile}{Survey Tile, subdivision of DR1 targets for survey processing,
tile168060 for example.}

\item{sourceid}{SuperWASP Object ID, 1SWASP J191645.46+474912.3 for example.}

\item{quiet}{If FALSE, suppress query message.}
}
\description{
The NASA Exoplanet Archive provides access to the public SuperWASP light
curves from Data Release One. See the SuperWASP page for more information
on the SuperWASP data products available at the NASA Exoplanet Archive.
The following table lists all of the data columns in the SuperWASP Time
Series table (superwasptimeseries) that can be returned through the
Exoplanet Archive's Application Programming Interface (API).
}
