% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoplanets-package.R
\docType{package}
\name{exoplanets-package}
\alias{exoplanets-package}
\alias{_PACKAGE}
\title{exoplanets: Access NASA's Exoplanet Archive Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The goal of exoplanets is to provide access to
    NASA's Exoplanet Archive TAP Service. For more information regarding
    the API please read the documentation
    <https://exoplanetarchive.ipac.caltech.edu/index.html>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/exoplanets/ (website) https://github.com/ropensci/exoplanets}
  \item Report bugs at \url{https://github.com/ropensci/exoplanets/issues}
}

}
\author{
\strong{Maintainer}: Tyler Littlefield \email{tylerlittlefield@hey.com} (\href{https://orcid.org/0000-0002-6020-1125}{ORCID})

Other contributors:
\itemize{
  \item Maëlle Salmon [contributor, reviewer]
}

}
\keyword{internal}
