% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoplanets.R
\docType{package}
\name{exoplanets}
\alias{exoplanets}
\alias{exoplanets-package}
\title{exoplanets Package}
\description{
This package provides tools for studying exoplantes
}
\author{
Samuel Macêdo \email{samuelmacedo@recife.ifpe.edu.br}
}
