% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{forget_exoplanets}
\alias{forget_exoplanets}
\title{Clear the exoplanets cache}
\usage{
forget_exoplanets()
}
\description{
Forget past results and reset the \code{exoplanets} cache.
}
\examples{
if (interactive()) {
  system.time(exoplanets("k2names"))
  system.time(exoplanets("k2names"))
  forget_exoplanets()
  system.time(exoplanets("k2names"))
}

}
