% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightcurves_data_dl.R
\name{lightcurves_data}
\alias{lightcurves_data}
\title{lightcurves_data_dl}
\usage{
lightcurves_data(kepid_table)
}
\arguments{
\item{nea_table}{data frame containing a column name kepid}
}
\description{
Download the lightcurve data according to the kepid provided
}
