% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_table.R
\name{nea_read_table}
\alias{nea_read_table}
\title{nea_read_table}
\usage{
nea_read_table(table, folder = NULL)
}
\arguments{
\item{table}{A string containing the table name. For the available tables check 'details' below.}

\item{folder}{If NULL (default) this function will search for
the folder created by nea_table(). If the data is in a different directory,
pass the PATH_TO_YOUR_FOLDER as a string.}
}
\value{
A data frame containing the data
}
\description{
This is a simple function for read the tables downloaded
by NASA Exoplanet Archive. You can use read.csv() function if
you need to be more especific about your data.
}
\details{
Use before nea_table() for download the data that you want
to work. You can check the available tables in
https://exoplanetarchive.ipac.caltech.edu/docs/program_interfaces.html#data.
}
\examples{
nea_table("q1_q17_dr24_tce") # Download your data
nea_read_table("q1_q17_dr24_tce")

}
