% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_table.R
\name{nea_table}
\alias{nea_table}
\title{nea_table}
\usage{
nea_table(table, columns = NULL, format = "csv", data_folder = TRUE,
  force = FALSE)
}
\arguments{
\item{table}{The table name (string). For the tables available
check details below.}

\item{columns}{Specify the columns that you want to download in the table.}

\item{format}{The archive extension. "csv" (default), "json", "xml", "ascii",
"ipac", "bar" and "pipe".}

\item{data_folder}{Create a separate folder for you data. If FALSE, the
data will be download in your root.}

\item{force}{Check if the data already exists. If TRUE, the data will
download and replace.}
}
\description{
A wrapper function for download any table in NASA Exoplanets Archive
}
\details{

}
