% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tableinfo}
\alias{tableinfo}
\title{Table Information}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 394 rows and 13 columns.
}
\source{
https://exoplanetarchive.ipac.caltech.edu/docs/TAP/usingTAP.html
}
\usage{
tableinfo
}
\description{
This dataset provides table information for NASA's Exoplanet
Archive TAP service. In particular, the table name, columns, and column
descriptions are provided.
}
\examples{
tableinfo

}
\keyword{datasets}
