% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exo.R
\name{exo}
\alias{exo}
\title{Access NASA's Exoplanet Archive by Table}
\usage{
exo(table = "exoplanets", cols = "default", quiet = FALSE)
}
\arguments{
\item{table}{The name of the table, see \code{exo_tables} for all available table
names.}

\item{cols}{Either "default" for default columns, "all" for all columns or
individual column names.}

\item{quiet}{If FALSE, suppress query message.}
}
\value{
A \code{data.frame} containing data for the respective table
}
\description{
Access NASA's Exoplanet Archive by table name. For a list of
all available tables, see \code{exo_tables}.
}
