% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoplanets.R
\name{exo}
\alias{exo}
\title{Access NASA's Exoplanet Archive by Table}
\usage{
exo(table = "exoplanets", cols = "default")
}
\arguments{
\item{table}{The name of the table, see \code{tbls} for all available table
names.}

\item{cols}{Either "default" for default columns, "all" for all columns or
individual column names separated by a comma, defaults to "default".
default set, defaults to TRUE.}
}
\description{
Access NASA's Exoplanet Archive by table name. For a list of
all available tables, see \code{tbls}.
}
\examples{
exoplanets <- exo("exoplanets")
k2candidates <- exo("k2candidates")
}
