% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoplanets.R
\name{exo_column_names}
\alias{exo_column_names}
\title{Access column names for each table}
\usage{
exo_column_names(table = "exoplanets", cols = "default")
}
\arguments{
\item{table}{The name of the table, see \code{tbls} for all available table
names.}

\item{cols}{Either "default" for default columns, "all" for all columns,
defaults to "default".}
}
\description{
Simply pull column names for a specified table. You can either
pull the default columns assigned to a table or all columns.
}
\examples{
str(exo_column_names("cumulative", "default"))
}
