% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoplanets.R
\name{exo_raw}
\alias{exo_raw}
\title{Access NASA's Exoplanet Archive by URL}
\usage{
exo_raw(query)
}
\arguments{
\item{query}{The full query URL. At this time, the only supported format
is CSV. If utilizing SQL, do not worry about spaces or single quotes, these
are escaped automatically.}
}
\value{
A \code{data.frame} containing data for the respective table
}
\description{
Access NASA's Exoplanet Archive by URL. For a list of example
queries, see the documentation,
\url{https://exoplanetarchive.ipac.caltech.edu/docs/program_interfaces.html}.
}
\examples{
qry <- "https://exoplanetarchive.ipac.caltech.edu/cgi-bin/nstedAPI/nph-nstedAPI?table=exoplanets"
exoplanets <- exo_raw(qry)
}
