% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoplanets.R
\name{exo_summary}
\alias{exo_summary}
\title{Exoplanet Summary}
\usage{
exo_summary(output = "list")
}
\arguments{
\item{output}{One of \code{list} or \code{dataframe}, defaults to list.}
}
\value{
An object of list or data.frame.
}
\description{
A list of data summarising the exoplanet database.
}
