% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoplanets.R
\name{exo_summary}
\alias{exo_summary}
\title{Exoplanet Summary}
\usage{
exo_summary(output = "list", progress = TRUE)
}
\arguments{
\item{output}{One of \code{list} or \code{dataframe}, defaults to list.}

\item{progress}{If FALSE, suppresses progress of request.}
}
\value{
An object of class \code{list} or \code{data.frame}.
}
\description{
A list of data summarising the exoplanet database. The goal of
this function is to provide the information found here
<\url{https://exoplanetarchive.ipac.caltech.edu/docs/counts_detail.html}>. By
default, a list is returned but a \code{data.frame} can be returned with
\code{output = "dataframe"}.
}
