% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoplanets.R
\name{exoplanets}
\alias{exoplanets}
\title{Retrieve Data from NASAs Exoplanet Archive}
\usage{
exoplanets(table, columns = NULL, format = "csv", progress = TRUE)
}
\arguments{
\item{table}{A table name}

\item{columns}{A vector of valid column names, by default will return all columns}

\item{format}{Desired format, either csv, tsv, or json}

\item{progress}{Whether or not to display the progress of the request}
}
\description{
A simple interface for accessing exoplanet data. At the bare minimum, a table
name is required. Tables names are documented in the `tap` dataset.
}
