% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_table.R
\name{nea_table}
\alias{nea_table}
\title{nea_table}
\usage{
nea_table(table, columns = "default", format = "csv", data_folder = TRUE,
  force = FALSE)
}
\arguments{
\item{table}{A string containing the table name. For the available tables check 'details' below.}

\item{columns}{Specify the columns that you want to download in the table. Define the columns
in a vector, for example, c("col1", "col2", "col3"). Use '"all" to download all available columns
and "default" for the default columns, see more in 'details'.}

\item{format}{The archive extension. "csv" (default), "json", "xml", "ascii", "ipac", "bar"
and "pipe".}

\item{data_folder}{Create a separate folder for you data. If FALSE, the
data will be download in your root.}

\item{force}{Check if the data already exists. If TRUE, the data will
download and replace.}
}
\description{
A wrapper function for download any table in NASA Exoplanets Archive
}
\details{
You can check the available tables in
https://exoplanetarchive.ipac.caltech.edu/docs/program_interfaces.html#data.
In this link you can also check the table documentation and there will be specified all
available columns and what are the 'default columns'.
}
\examples{
# download default columns for the exoplanets table
nea_table("exoplanets")

# download specific columns
nea_table("q1_q17_dr24_tce", columns = c("kepid", "av_training_set"))

}
